refactor(worker): reorganize source tree
