+++ apps/worker/package.json
@@ -4,8 +4,9 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "start": "bun run index.ts",
-    "dev": "bun run index.ts"
+    "start": "bun run src/index.ts",
+    "dev": "bun run src/index.ts",
+    "dev:test": "bun run src/dev/test.ts"
   },
   "devDependencies": {
     "@types/bun": "latest",
