+++ apps/worker/src/utils/mailparser.ts
