+++ apps/worker/src/utils/trcp.ts
@@ -1,5 +1,5 @@
 import { createTRPCClient, httpBatchLink } from "@trpc/client";
-import type { AppRouter } from "../../server/src/routers";
+import type { AppRouter } from "../../../server/src/routers";
 
 export const trpcClient = createTRPCClient<AppRouter>({
   links: [
